% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_scale}
\alias{add_scale}
\title{Adding a scale bar to the image}
\usage{
add_scale(
  image,
  resolution,
  downsample,
  margins = c(0.1, 0.1),
  width = 0.2,
  outputFile = NULL
)
}
\arguments{
\item{image}{a file path as a character or a magick-image}

\item{resolution}{Image resolution as acquired via \code{\link{getSectionImage}}}

\item{margins}{Percent ratios as distance from the bottom left corner of the image}

\item{width}{Percent ratio of image width to use for scale width}

\item{outputFile}{file path of the output.}
}
\description{
Adding a scale bar to the image
}
