% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{getAllExperiments}
\alias{getAllExperiments}
\title{Acquire a table listing all experiments}
\usage{
getAllExperiments(planeOfSection)
}
\arguments{
\item{planeOfSection}{coronal or sagittal}
}
\value{
Table of experiments with gene_symbol, entrez_id, experiment_id, specimen_id
}
\description{
Acquire a table listing all experiments
}
