% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{getAtlasImage}
\alias{getAtlasImage}
\title{Get atlas image data for a given ID}
\usage{
getAtlasImage(imageID)
}
\arguments{
\item{imageID}{id of the image}
}
\value{
A list containing all information about the Atlas Image
}
\description{
Get atlas image data for a given ID
}
