% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{getGeneDataset}
\alias{getGeneDataset}
\title{Acquire datasets for genes}
\usage{
getGeneDataset(gene, planeOfSection = c("sagittal", "coronal"))
}
\arguments{
\item{gene}{A mouse gene symbol}

\item{planeOfSection}{coronal or sagittal}
}
\value{
Dataset ID of the first section dataset matching the description
}
\description{
Acquire datasets for genes
}

