% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{getGeneDatasets}
\alias{getGeneDatasets}
\title{Acquire datasets for genes}
\usage{
getGeneDatasets(gene, planeOfSection = c("sagittal", "coronal", "all"))
}
\arguments{
\item{gene}{A mouse gene symbol}

\item{planeOfSection}{coronal or sagittal}
}
\value{
Dataset ID of section datasets matching the description (gene and section)
}
\description{
Acquire datasets for genes
}
