% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{getImageID}
\alias{getImageID}
\title{Acquire image id centered on the region}
\usage{
getImageID(datasetID, regionID)
}
\value{
A named list including the ID of the image and coordinates of the brain region
}
\description{
Acquire image id centered on the region
}
