% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{getSectionImage}
\alias{getSectionImage}
\title{Get section image data for a given ID}
\usage{
getSectionImage(imageID)
}
\arguments{
\item{imageID}{id of the image}
}
\value{
A list containing all information about Section Image
}
\description{
Get section image data for a given ID
}
