% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{getStructureIDs}
\alias{getStructureIDs}
\title{Acquire a list of structure ids}
\usage{
getStructureIDs()
}
\value{
A data frame with 2 columns. All names are in lower case
}
\description{
Acquire a list of structure ids
}
