% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSync.R
\name{imageToImage}
\alias{imageToImage}
\title{For a list of target SectionDataSets, find the closest SectionImage and (x,y) location as defined by a seed SectionImage and seed (x,y) pixel location.}
\usage{
imageToImage(seedImage, x, y, datasets)
}
\description{
For a list of target SectionDataSets, find the closest SectionImage and (x,y) location as defined by a seed SectionImage and seed (x,y) pixel location.
}
