% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{imageToImage2D}
\alias{imageToImage2D}
\title{Finds the closest points in images provided in the seed image and seed coordinates}
\usage{
imageToImage2D(seedImage, x, y, images)
}
\description{
Finds the closest points in images provided in the seed image and seed coordinates
}
