% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSync.R
\name{structureToImage}
\alias{structureToImage}
\title{Acquire image id centered on the region}
\usage{
structureToImage(datasetID, regionIDs)
}
\value{
A named list including the ID of the image and coordinates of the brain region
}
\description{
Acquire image id centered on the region
}
