% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{centerImage}
\alias{centerImage}
\title{Center and crop the image}
\usage{
centerImage(imageFile, x, y, xProportions = c(0.1, 0.1),
  yProportions = c(0.1, 0.1), outputFile)
}
\description{
Center and crop the image
}

