% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenBrain.R
\name{getStructureIDs}
\alias{getStructureIDs}
\title{Acquire a list of structure ids}
\usage{
getStructureIDs(
  ontologyName = c("Mouse Brain Atlas", "Developing Mouse Brain Atlas",
    "Human Brain Atlas", "Developing Human Brain Atlas", "Non-Human Primate Brain Atlas",
    "Glioblastoma"),
  graphID = NULL
)
}
\arguments{
\item{ontologyName}{name of the ontology to download}

\item{graphID}{alternative to ontologyName, graph IDs can be provided to acquire the corresponding ontology. This will override the ontologyName provided.}
}
\value{
A data frame with 4 columns: id, names, acronyms and parent
}
\description{
This function acquires the list of structure ids and names that correspond to the structure ids in Allen Brain Atlas.
Default behavior is to get the mouse brain atlas ontology. Most downstream functions in this package are specific to mouse brain ontology
}
