% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSync.R
\name{imageToImage2D}
\alias{imageToImage2D}
\title{For a list of target SectionImages, find the closest (x,y) location as defined by a seed SectionImage and seed (x,y) location.}
\usage{
imageToImage2D(seedImage, x, y, images)
}
\description{
For a list of target SectionImages, find the closest (x,y) location as defined by a seed SectionImage and seed (x,y) location.
}
